Create bullet item widget.
