+++ lib/features/settings/presentation/widgets/server_setup_instructions.dart
@@ -1,42 +1,34 @@
 import 'package:flutter/material.dart';
 
+import '../../../../core/widgets/bullet_item.dart';
+
 class ServerSetupInstructions extends StatelessWidget {
-  const ServerSetupInstructions({Key key}) : super(key: key);
+  final bool showWarning;
+
+  const ServerSetupInstructions({
+    Key key,
+    this.showWarning = true,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
-      children: const <Widget>[
-        Padding(
-          padding: EdgeInsets.only(bottom: 10),
-          child: Text(
-            'This app is not registered to a Tautulli server.',
-            style: TextStyle(fontWeight: FontWeight.bold),
+      children: <Widget>[
+        if (showWarning)
+          const Padding(
+            padding: EdgeInsets.only(bottom: 10),
+            child: Text(
+              'This app is not registered to a Tautulli server.',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
           ),
-        ),
-        Text('To register with a Tautulli server:'),
-        BulletItem('Open the Tautulli web interface on another device'),
-        BulletItem('Navigate to Settings > Tautulli Remote App'),
-        BulletItem('Select \'Register a new device\''),
-        BulletItem('Use the button below to add a new server'),
+        const Text('To register with a Tautulli server:'),
+        const BulletItem('Open the Tautulli web interface on another device'),
+        const BulletItem('Navigate to Settings > Tautulli Remote App'),
+        const BulletItem('Select \'Register a new device\''),
+        const BulletItem('Use the below option to register with a new server'),
       ],
-    );
-  }
-}
-
-class BulletItem extends StatelessWidget {
-  final String text;
-
-  const BulletItem(this.text, {Key key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(
-        left: 10,
-      ),
-      child: Text('• $text'),
     );
   }
 }
